% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockchain.R
\name{apis}
\alias{apis}
\title{Available APIs}
\usage{
apis()
}
\value{
a data frame.
}
\description{
Get available APIs.
}
\details{
This function returns a data frame containing a column of APIs and logical columns indicating whether an API is available for ticker
information or for wallet information.
}
\examples{
apis()
}
