% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coin_helpers}
\alias{coin_helpers}
\alias{coin_id}
\alias{coin_name}
\alias{coin_symbol}
\title{Coin helper functions}
\usage{
coin_id(x)

coin_name(x)

coin_symbol(x)
}
\arguments{
\item{x}{a coin id, name or ticker symbol.}
}
\value{
a vector the same length as \code{x}.
}
\description{
These helpers return ID, name or ticker symbol for cryptocurrencies based on either of the other two.
Symbol inputs are not case sensitive but IDs and names are.
}
\examples{
coin_id(c("btc", "Ethereum"))
coin_name(c("BTC", "BCH"))
coin_symbol("Bitcoin")
}
