% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{coin_value}
\alias{coin_value}
\title{Get the current value of a cryptocurrency}
\usage{
coin_value(crypto = "BTC", fiat)
}
\arguments{
\item{crypto}{character, cryptocurrency trading symbol, e.g. BTC.}

\item{fiat}{character, fiat currency trading symbol, e.g., USD.}
}
\value{
a data frame.
}
\description{
Get the current value of cryptocurrency in terms of fiat currencies.
}
\details{
The current implementation of this function only accepts code{crypto = "BTC"}.
\code{crypto} and \code{fiat} are not case sensitive.
\code{fiat} may be a vector. If missing, all available fiat currency values are returned.
}
\examples{
coin_value()
}
