% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coins}
\alias{coins}
\title{Avaliable cryptocurrencies.}
\format{A data frame with 1234 rows and 3 columns: cryptocurrency \code{id}, \code{name} and ticker \code{symbol}.
This data frame provides the list of available cryptocurrenices at the time of package publication.
In case this table is out of date, the full current list can always be obtained from the first three columns returned by \link{code{ticker}}.}
\usage{
coins
}
\description{
Available cryptocurrencies, Bitcoin and altcoins, on \code{coinmarketcap.com}.
}
\keyword{datasets}
