% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btc_conversion.R
\name{convert_btc}
\alias{convert_btc}
\alias{satoshi}
\alias{bits}
\alias{mbtc}
\title{Convert Bitcoin units}
\usage{
convert_btc(x, from = "BTC", to = "bits")

satoshi(x)

bits(x)

mbtc(x)
}
\arguments{
\item{x}{a numeric value.}

\item{from}{the units of \code{x}.}

\item{to}{new units.}
}
\value{
numeric value.
}
\description{
Convert among different units of measure for Bitcoin.
}
\details{
This complete conversion function, \code{convert_btc}, converts a value \code{x} among the following units: BTC, mBTC, bits and satoshi.
These are the recommended ways of specifying the \code{from} and \code{to} arguments, though BTC and mBTC
may be passed as btc and mbtc, respectively. BTC may also be written out as Bitcoin/bitcoin and will also be internally converted to BTC.

Three shorthand functions, \code{mBTC}, \code{bits} and \code{satoshi}, are also available if you forget the conversion.
These assume the input is Bitcoin and multiply your value straight to mBTC, bits or satoshi, respectively. See examples.
}
\examples{
convert_btc(1, "BTC", "mBTC")
convert_btc(1, "BTC", "bits")
convert_btc(1, "BTC", "satoshi")
convert_btc(0.001, "mBTC", "bits")
convert_btc(1, "mBTC", "bitcoin")
convert_btc(100, "satoshi", "bits")
satoshi(1e-8)
bits(1e-6)
mbtc(0.001)
}
