% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockchain.R
\name{fiat_symbols}
\alias{fiat_symbols}
\title{Valid fiat ticker symbols}
\usage{
fiat_symbols(api = NULL)
}
\arguments{
\item{api}{character or \code{NULL}, the API. See \code{\link{apis}}.}
}
\value{
a data frame if \code{api = NULL}, otherwise a vector.
}
\description{
Get valid fiat currency ticker symbols for available ticker/market cap APIs.
}
\examples{
fiat_symbols()
fiat_symbols(api = "coinmarketcap.com")
fiat_symbols(api = "blockchain.info")
}
\seealso{
\code{\link{apis}}
}
