% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockchain.R
\docType{package}
\name{rockchain}
\alias{rockchain}
\alias{rockchain-package}
\title{rockchain: insert text here.}
\description{
The \code{rockchain} package provides a simple interface to \href{https://blockchain.info/}{blockchain.info} ticker and wallet data.
At this time the package is very basic. It does not offer any additional functionality beyond basic wallet transaction data retrieval
or integrate with other blockchain APIs. It currently only returns bitcoin data. Access to other cryptocurrencies in not yet implemented.
}
