% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Cryptocurrency ticker/market cap data}
\usage{
ticker(crypto = NULL, convert = "USD", api = "coinmarketcap.com")
}
\arguments{
\item{crypto}{character or \code{NULL}, cryptocurrency trading symbol. See details.}

\item{convert}{character or \code{NULL}, conversion currency trading symbol. See details.}
}
\value{
a data frame. Available columns depend on the selected API.
}
\description{
Get cryptocurrency ticker/market cap data from a specified API.
}
\details{
The current implementation of this function offers two common APIs: \code{coinmarketcap.com} and \code{blockchain.info}.
\code{coinmarketcap.com} returns ticker/market cap information for many different cryptocurrencies and their value
can be expressed in any of several fiat currencies (e.g., USD) or other cryptocurrencies (e.g. BTC).
\code{blockchain.info} is specific to Bitcoin. It only returns Bitcoin information, which can be expressed in any of several fiat currencies.

\code{crypto} and \code{convert} are not case sensitive. Both arguments are handled somewhat different for each API.
For simplicity, there are available wrapper functions,
\code{\link{cap}} and \code{\link{bcinfo}}, that help separate the APIs and what information you intend to obtain from each.

\code{crypto} and \code{convert} may be vectors, though of course for \code{blockchain.info}, \code{crypto} is simply ignored.
For \code{coinmarketcap.com}, the default \code{crypto = NULL} will return all available cryptocurrencies tracked by \code{coinmarketcap.com}
\code{convert} may be \code{NULL}. This is handled differently by each API. For \code{coinmarketcap.com}, USD is assumed in order to limit the number of API calls.
For \code{blockchain.info} all available fiat currency-Bitcoin ticker pairs are returned.
}
\examples{
\dontrun{
ticker()
ticker(crypto = c("ETH", "LTC"), convert = c("EUR", "GBP", "BTC"))
ticker(api = "blockchain.info")
}
}
\seealso{
\code{\link{cap}}, \code{\link{bcinfo}}
}
