% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet_helpers.R
\name{wallet_helpers}
\alias{wallet_helpers}
\alias{hash}
\alias{address}
\alias{received}
\alias{sent}
\alias{balance}
\alias{txn}
\alias{transactions}
\title{Wallet helper functions}
\usage{
hash(x)

address(x)

received(x)

sent(x)

balance(x)

txn(x)

transactions(x)
}
\arguments{
\item{x}{a wallet list object. See \link{wallet}.}
}
\value{
vectorized results pertaining to each element in \code{x}. For \code{transactions}, a combined data frame with a wallet \code{address} column.
}
\description{
These helpers are simple getter functions for convenient access to specific wallet information.
}
\examples{
\dontrun{
x <- wallet("1KennyH9grzif79WbaQDHpqgTnm25j4rRj")
hash(x)
address(x)
received(x)
sent(x)
balance(x)
txn(x)
transactions(x)
}
}
