% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockchain.R
\name{wallet_info}
\alias{wallet_info}
\title{Get wallet info}
\usage{
wallet_info(id)
}
\arguments{
\item{id}{character, wallet ID.}
}
\value{
a list.
}
\description{
Get wallet info from blockchain.info.
}
\details{
The current implementation of this function grabs information for specified wallets only from \url{blockchain.info}.
}
\examples{
\dontrun{wallet_info("1KennyH9grzif79WbaQDHpqgTnm25j4rRj")}
}
