% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wlt}
\alias{wlt}
\title{Unit test wallet data examples.}
\format{A list of seven wallet lists.}
\usage{
wlt
}
\description{
Example downloaded wallet data from the \code{coinmarketcap.com} API.
}
\details{
This data is only used for unit testing during package development and maintenance. It only exists so that unit tests can be run when the
\code{blockchain.info} API is being extremely fussy and constantly rejecting API calls. This is not a permanent solution. Eventually
the data set will be removed and, if necessary, an alternative, more dependable Bitcoin blockchain API will be used instead of \code{blockchain.info}.
}
\keyword{datasets}
