% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockchain.R
\docType{package}
\name{rockchain}
\alias{rockchain}
\alias{rockchain-package}
\title{rockchain: a blockchain package for R.}
\description{
Access blockchain and cryptocurrency data.
}
\details{
The \code{rockchain} package provides simple interfaces to \href{https://coinmarketcap.com/}{coinmarketcap.com} ticker and market cap data and
\href{https://blockchain.info/}{blockchain.info} ticker and Bitcoin wallet data.
The current package does not yet support additional blockchain APIs.
It does not offer functionality beyond cryptocurrency market data retrieval and Bitcoin wallet transaction data retrieval.
}
