% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{wallet}
\alias{wallet}
\title{Get wallet info}
\usage{
wallet(id, satoshi = FALSE, offset = 0, tx_max = 100)
}
\arguments{
\item{id}{character, wallet ID.}

\item{satoshi}{logical, if set to \code{TRUE}, retain Satoshi units from blockchain.info. Otherwise all values are in bitcoin (\code{Satoshis / 10e7}).}

\item{offset}{integer, shift transaction number retrieval forward by this amount.}

\item{tx_max}{integer, maximum number of transactions to return. Defaults to 100. \code{NULL} will attempt to retrieve all transactions.}
}
\value{
a list.
}
\description{
Get wallet info from blockchain.info.
}
\details{
The current implementation of this function grabs information for specified wallets only from \url{blockchain.info}.
}
\examples{
\dontrun{wallet("1KennyH9grzif79WbaQDHpqgTnm25j4rRj")}
}
